requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==4.2.9
django-admin-black==1.0.8
django-allauth==0.63.6
django-browser-reload==1.12.1
djangorestframework==3.15.2
gunicorn==21.2.0
idna==3.7
packaging==24.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
str2bool==1.1
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.6.0
